requirements.txt
gspread
oauth2client
python-dotenv
pandas
requests
Pillow
pyzbar
google-cloud-vision
streamlit